requirements.txt
Flask >= 3.0.3
Flask-SQLAlchemy >= 3.1.1
Flask-Login >= 0.6.3
Werkzeug >= 3.0.2
Flask_WTF >= 1.2.1
mysqlclient >= 2.2.4